Move the in-memory store to elide-core (#447)
