+++ elide-core/pom.xml
@@ -122,6 +122,11 @@
             <artifactId>jansi</artifactId>
             <version>1.14</version>
         </dependency>
+        <dependency>
+            <groupId>org.reflections</groupId>
+            <artifactId>reflections</artifactId>
+            <version>0.9.10</version>
+        </dependency>
 
         <!-- Test -->
         <dependency>
