+++ elide-datastore/elide-datastore-inmemorydb/src/main/java/com/yahoo/elide/datastores/inmemory/InMemoryDataStore.java
@@ -5,64 +5,13 @@
  */
 package com.yahoo.elide.datastores.inmemory;
 
-import com.yahoo.elide.core.DataStore;
-import com.yahoo.elide.core.DataStoreTransaction;
-import com.yahoo.elide.core.EntityDictionary;
-
-import org.reflections.Reflections;
-import org.reflections.scanners.SubTypesScanner;
-import org.reflections.scanners.TypeAnnotationsScanner;
-import org.reflections.util.ClasspathHelper;
-import org.reflections.util.ConfigurationBuilder;
-
-import lombok.Getter;
-
-import java.util.Map.Entry;
-import java.util.concurrent.ConcurrentHashMap;
-
-import javax.persistence.Entity;
-
 /**
  * Simple non-persistent in-memory database.
+ * @deprecated Use {@link com.yahoo.elide.core.datastore.inmemory.InMemoryDataStore}
  */
-public class InMemoryDataStore implements DataStore {
-    private final ConcurrentHashMap<Class<?>, ConcurrentHashMap<String, Object>> dataStore =
-            new ConcurrentHashMap<>();
-    @Getter private EntityDictionary dictionary;
-    @Getter private final Package beanPackage;
-
+@Deprecated
+public class InMemoryDataStore extends com.yahoo.elide.core.datastore.inmemory.InMemoryDataStore {
     public InMemoryDataStore(Package beanPackage) {
-        this.beanPackage = beanPackage;
-    }
-
-    @Override
-    public void populateEntityDictionary(EntityDictionary dictionary) {
-        Reflections reflections = new Reflections(new ConfigurationBuilder()
-                .addUrls(ClasspathHelper.forPackage(beanPackage.getName()))
-                .setScanners(new SubTypesScanner(), new TypeAnnotationsScanner()));
-        reflections.getTypesAnnotatedWith(Entity.class).stream()
-                .filter(entityAnnotatedClass -> entityAnnotatedClass.getPackage().getName()
-                        .startsWith(beanPackage.getName()))
-                .forEach(dictionary::bindEntity);
-        this.dictionary = dictionary;
-    }
-
-    @Override
-    public DataStoreTransaction beginTransaction() {
-        return new InMemoryTransaction(dataStore, dictionary);
-    }
-
-    @Override
-    public String toString() {
-        StringBuilder sb = new StringBuilder();
-        sb.append("Data store contents ");
-        for (Class<?> cls : dataStore.keySet()) {
-            sb.append("\n Table ").append(cls).append(" contents \n");
-            ConcurrentHashMap<String, Object> data = dataStore.get(cls);
-            for (Entry<String, Object> e : data.entrySet()) {
-                sb.append(" Id: ").append(e.getKey()).append(" Value: ").append(e.getValue());
-            }
-        }
-        return sb.toString();
+        super(beanPackage);
     }
 }
